Created Nested Workflows
Created nested workflows to run staged relaxations into static energy calculations. Also switched from pbe to pbesol for population analysis workflows.
