+++ src/warrenapp/workflows/population_analysis/badelf_hse.py
@@ -3,36 +3,12 @@
 # from warrenapp.workflows.population_analysis.badelf_pbe import (
 #     StaticEnergy__Warren__PrebadelfPbe,
 # )
-from warrenapp.workflows.population_analysis.base import (
-    VaspBadElfBase,
-    prebadelf_incar_settings,
-)
+from warrenapp.workflows.population_analysis.base import VaspBadElfBase
 
 # from warrenapp.workflows.static_energy import StaticEnergy__Warren__SeededHse
-from warrenapp.workflows.static_energy.hse import StaticEnergy__Warren__Hse
-
-
-class StaticEnergy__Warren__PrebadelfHse(StaticEnergy__Warren__Hse):
-    """
-    Runs a static energy calculation with a high-density FFT grid under HSE
-    settings from the Warren Lab. Results can be used for Bader analysis where
-    the ELF is used as the reference instead of the CHGCAR.
-
-    We do NOT recommend running this calculation on its own. Instead, you should
-    use the full workflow, which runs this calculation AND the following bader
-    analysis for you. This S3Task is only the first step of that workflow.
-    """
-
-    # The key thing for bader analysis is that we need a very fine FFT mesh. Other
-    # than that, it's the same as a static energy calculation.
-    incar = StaticEnergy__Warren__Hse.incar.copy()
-    incar.update(prebadelf_incar_settings)
-
-
-# We prefer to use a workflow that seeds the HSE calculation with a PBE calculation
-# class StaticEnergy__Warren__PrebadelfSeededHse(StaticEnergy__Warren__SeededHse):
-#     initial_calculation = StaticEnergy__Warren__PrebadelfPbe
-#     second_calculation = StaticEnergy__Warren__PrebadelfHse
+from warrenapp.workflows.population_analysis.prebader_badelf_dft import (
+    StaticEnergy__Warren__PrebadelfHse,
+)
 
 
 class PopulationAnalysis__Warren__BadelfHse(VaspBadElfBase):
