+++ src/warrenapp/workflows/population_analysis/bader_badelf_pbesol.py
@@ -1,18 +1,19 @@
 # -*- coding: utf-8 -*-
 
+
 from simmate.engine.workflow import Workflow
 
 from warrenapp.workflows.population_analysis.base import VaspBaderBadElfBase
 from warrenapp.workflows.population_analysis.prebader_badelf_dft import (
-    StaticEnergy__Warren__PrebadelfHse,
+    StaticEnergy__Warren__PrebadelfPbesol,
 )
 
 
-class PopulationAnalysis__Warren__BaderBadelfHse(VaspBaderBadElfBase):
+class PopulationAnalysis__Warren__BaderBadelfPbesol(VaspBaderBadElfBase):
     """
     Runs a static energy calculation using an extra-fine FFT grid using vasp
     and then carries out Badelf and Bader analysis on the resulting charge density.
-    Uses the Warren lab settings HSE settings.
+    Uses the Warren lab settings Pbesol settings.
     """
 
-    static_energy_prebadelf: Workflow = StaticEnergy__Warren__PrebadelfHse
+    static_energy_prebadelf: Workflow = StaticEnergy__Warren__PrebadelfPbesol
