+++ src/warrenapp/workflows/population_analysis/base.py
@@ -72,8 +72,8 @@ class PopulationAnalysis__Warren__BaderEmpty(S3Workflow):
             check_required_files(directory=directory, required_files=required_files)
         except:
             get_density_file_empty(
-                directory=directory, 
-                structure=structure, 
+                directory=directory,
+                structure=structure,
                 analysis_type=analysis_type,
                 min_charge=min_charge,
             )
@@ -119,8 +119,8 @@ class PopulationAnalysis__Warren__GetAtomChgcar(S3Workflow):
             check_required_files(directory=directory, required_files=required_files)
         except:
             get_density_file_empty(
-                directory=directory, 
-                structure=structure, 
+                directory=directory,
+                structure=structure,
                 analysis_type=analysis_type,
                 min_charge=min_charge,
             )
@@ -171,7 +171,7 @@ class VaspBaderBase(Workflow):
             ).result()
             # Find electride sites, place empty atoms, and run bader
             PopulationAnalysis__Warren__BaderEmpty.run(
-                structure=structure, 
+                structure=structure,
                 directory=prebader_result.directory,
                 min_charge=min_charge,
             )
@@ -238,7 +238,7 @@ class VaspBadElfBase(Workflow):
             # Find electride sites, place empty atoms, and generate chgcar
             # like files for each atomic site
             PopulationAnalysis__Warren__GetAtomChgcar.run(
-                directory=prebadelf_result.directory, 
+                directory=prebadelf_result.directory,
                 analysis_type="badelf",
                 min_charge=min_charge,
             )
@@ -318,7 +318,7 @@ class VaspBaderBadElfBase(Workflow):
             # Find electride sites, place empty atoms, and get atom charges
             # in CHGCAR format
             PopulationAnalysis__Warren__GetAtomChgcar.run(
-                directory=directory, 
+                directory=directory,
                 analysis_type="both",
                 min_charge=min_charge,
             ).result()
@@ -336,7 +336,7 @@ class VaspBaderBadElfBase(Workflow):
             # )
             # Run bader analysis
             PopulationAnalysis__Warren__BaderEmpty.run(
-                structure=structure, 
+                structure=structure,
                 directory=directory,
                 min_charge=min_charge,
             ).result()
@@ -378,23 +378,3 @@ class VaspBaderBadElfBase(Workflow):
             # Copy bader files
             for file in bader_files:
                 shutil.copy(directory / file, directory / "bader")
-
-
-# We want to define the settings that will be used when updating static energy
-# workflows for prebader or prebadelf DFT calculations. We do that here so
-# that we don't need to do it in every inheriting class.
-prebader_incar_settings = dict(
-    NGXF__density_a=10,
-    NGYF__density_b=10,
-    NGZF__density_c=10,
-    LAECHG=True,  # write core charge density to AECCAR0 and valence to AECCAR2
-)
-prebadelf_incar_settings = dict(
-    NGX__density_a=10,  # Note that these set the FFT grid while the pre-Bader task sets the
-    NGY__density_b=10,  # fine FFT grid (e.g. useds NGX instead of NGXF)
-    NGZ__density_c=10,
-    LELF=True,  # Writes the ELFCAR
-    NPAR=1,  # Must be set if LELF is set
-    PREC="Single",  # ensures CHGCAR grid matches ELFCAR grid
-    LAECHG=True,  # write core charge density to AECCAR0 and valence to AECCAR2
-)
