+++ src/warrenapp/workflows/static_energy/hsesol.py
@@ -3,6 +3,7 @@
 from warrenapp.workflows.relaxation.hsesol import Relaxation__Warren__Hsesol
 from warrenapp.workflows.static_energy.hse import hse_static_settings
 
+
 class StaticEnergy__Warren__Hsesol(Relaxation__Warren__Hsesol):
     """
     Performs a static energy calculation based on the settings for Warren Lab
